✨ ShoppingListのParallel Routesの対応
