+++ src/app/(root)/shopping-list/@login/page.tsx
@@ -1,0 +1,11 @@
+import SuggestLogin from "@/src/components/suggest-login";
+
+const page = async () => {
+  return (
+    <>
+      <SuggestLogin page="ShoppingList" />
+    </>
+  );
+};
+
+export default page;
