+++ src/app/(root)/shopping-list/@view/page.tsx
@@ -3,8 +3,8 @@ import { getMyCartList } from "@/src/actions/getMyCartList";
 
 import TopBar from "@/src/components/layout/top-bar";
 
-import { CartListItemForm, CartListItemFormValues } from "./_components/cart-list-item-form";
-import { MemoForm, MemoFormValues } from "./_components/memo-form";
+import { CartListItemForm, CartListItemFormValues } from "../_components/cart-list-item-form";
+import { MemoForm, MemoFormValues } from "../_components/memo-form";
 
 const page = async () => {
   const memos = await getMemos();
