+++ src/app/(root)/shopping-list/_components/cart-list-item-form/index.ts
